ref: Splits the RawPacketCache/CachingTransformer.
